feat: add post detail page testing (not 100% yet)
